requirements.txt
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
decorator==4.4.2
Flask==1.1.2
future==0.18.2
google-auth==1.14.0
gspread==3.4.2
gunicorn==20.0.4
httplib2==0.17.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==12.6.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
